backend/requirements.txt
django~=4.1.0
djangorestframework~=3.14.0
requests